fixed names of projects and files
