+++ Ex04.Menus.Test/CountLowerCaseHandler.cs
