+++ Ex04.Menus.Test/MenuManager.cs
