+++ Ex04.Menus.Test/ShowDateHandler.cs
