+++ Ex04.Menus.Test/ShowTimeHandler.cs
